make the 2.2 branch 64-bit clean as well
2005-08-30  Paolo Bonzini  <bonzini@gnu.org>

	* libgst/oop.c: Fix pasto (SIZEOF_LONG instead of SIZEOF_OOP).
	* libgst/comp.c: Do not emit PUSH_INTEGER for integers equal or
	bigger than 2^31.
	* libgst/byte.h: Change final argument of compile_byte to int.
	* libgst/byte.c: Likewise.

git-archimport-id: bonzini@gnu.org--2004b/smalltalk--devo--2.2--patch-58

+++ libgst/byte.c
@@ -79,7 +79,7 @@ _gst_line_number (int n, mst_Boolean force)
 }
 
 void
-_gst_compile_byte (gst_uchar byte, long arg)
+_gst_compile_byte (gst_uchar byte, int arg)
 {
   int num_bytes;
   long n;
